Add files via upload
Sorry I didn't do this sooner, but I have had terrible experiences with git. I am adding the admin.py which will allow us to display models to the admin interface. Currently supports seeing a data file model which represents the file we are trying to upload to the admin page. You CANNOT save a file if you try to it will crash the web app.